# public/api-docs/models/openapi.yaml
openapi: 3.1.1
info:
  title: '🏛️ MiW REST api'
  description: '[UPM] MiW REST api'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 2.1.5
  contact:
    name: Máster Universitario en Ingeniería Web
    url: https://miw.etsisi.upm.es/
    email: miw.etsisi@upm.es
servers:
  - url: '{schema}://{servername}:{port}/api/v1'
    description: 'MiW Production server (uses live data)'
    variables:
      schema:
        enum:
          - http
          - https
        default: https
      servername:
        default: 127.0.0.1
      port:
        default: '8000'
...
paths:
...
  /results/{resultId}.{format}:
    get:
      summary: "Retrieve a single Result by its ID with format especified ouput."
      description: |
        - Returns the requested Result if the user has permission (ROLE_ADMIN can see any; ROLE_USER can only see their own).
        - Output can be in JSON or XML format.
        - Includes ETag for cache/conditional requests (304 Not Modified).
      operationId: "api_results_get_format"
      tags:
        - "🔢 Results"
      security:
        - MiWApiSecurity: []
      parameters:
        - $ref: '#/components/parameters/resultId'
        - $ref: '#/components/parameters/formatParam'
        - $ref: '#/components/parameters/ifNoneMatchParam'
      responses:
        "200":
          description: "OK - Returns the requested Result."
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
            application/xml:
              schema:
                $ref: '#/components/schemas/Result'
              example: |
                <?xml version="1.0" encoding="UTF-8"?>
                <result>
                  <entry xmlns:atom="http://www.w3.org/2005/Atom" id="1">
                    <result>111</result>
                    <time><![CDATA[2025-01-01 10:10:10]]</time>
                    <user xmlns:atom="http://www.w3.org/2005/Atom" id="1508">
                      <email>miw.etsisi@upm.es</email>
                      <roles>
                        <role>ROLE_USER</role>
                        <role>ROLE_ADMIN</role>
                      </roles>
                      <link rel="parent" href="/api/v1/users"/>
                      <link rel="self" href="/api/v1/users/1508"/>
                    </user>
                    <link rel="parent" href="/api/v1/results"/>
                    <link rel="self" href="/api/v1/results/1"/>
                  </entry>
                </result>
        "304":
          $ref: "#/components/responses/304_NotModified_Response"
        "401":
          $ref: "#/components/responses/401_Unauthorized_Response"
        "403":
          $ref: "#/components/responses/403_Forbidden_NotOwn_Response"
        "404":
          $ref: "#/components/responses/404_Resource_Not_Found_Response"
    put:
      tags: [ '🔢 Results' ]
      summary: Updates the Result resource with format.
      description: 'Updates the result identified by `resultId`.Output can be in JSON or XML format.'
      operationId: api_results_put_format
      parameters:
        - $ref: '#/components/parameters/resultId'
        - $ref: '#/components/parameters/formatParam'
        - $ref: '#/components/parameters/ifMatchParam'
      security:
        - MiWApiSecurity: [ ]
      requestBody:
        $ref: '#/components/requestBodies/ResultRequestBodyPut'
      responses:
        '209':
          description: '`Content Returned`: result previously existed and is now updated'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
            application/xml:
              schema:
                $ref: '#/components/schemas/Result'
              example: |
                <?xml version="1.0" encoding="UTF-8"?>
                <result>
                  <entry xmlns:atom="http://www.w3.org/2005/Atom" id="1">          
                    <result>111</result>                        
                    <time><![CDATA[2023-12-12 10:10:10]]</time>                    
                    <user xmlns:atom="http://www.w3.org/2005/Atom" id="1508">
                      <email>miw.etsisi@upm.es</email>
                      <roles>
                        <role>ROLE_USER</role>
                        <role>ROLE_ADMIN</role>
                      </roles>                    
                      <link rel="parent" href="/api/v1/users"/>
                      <link rel="self" href="/api/v1/users/1508"/>
                    </user>                    
                    <link rel="parent" href="/api/v1/results"/>
                    <link rel="self" href="/api/v1/results/1"/>
                  </entry>
                </result>
        '401':
          $ref: '#/components/responses/401_Unauthorized_Response'
        '403':
          $ref: '#/components/responses/403_Forbidden_NotOwn_Response'
        '404':
          $ref: '#/components/responses/404_Resource_Not_Found_Response_Put_Result'
        '412':
          $ref: '#/components/responses/412_Precondition_Failed_Response'
        '422':
          $ref: '#/components/responses/422_UnprocessableEntity_Response_Result'
    delete:
      tags: [  '🔢 Results' ]
      summary: Removes the Result resource with format.
      description: 'Deletes the result identified by `resultId`. Output can be in JSON or XML format..'
      operationId: api_results_delete_format
      parameters:
        - $ref: '#/components/parameters/resultId'
        - $ref: '#/components/parameters/formatParam'
      security:
        - MiWApiSecurity: [ ]
      responses:
        '204':
          description: 'Result deleted &lt;Response body is empty&gt;'
        '401':
          $ref: '#/components/responses/401_Unauthorized_Response'
        '403':
          $ref: '#/components/responses/403_Forbidden_NotOwn_Response_json'
        '404':
          $ref: '#/components/responses/404_Resource_Not_Found_Response'
    options:
      tags: [ '🔢 Results' ]
      summary: Provides the list of HTTP supported methods.
      description: 'Return a `Allow` header with a comma separated list of HTTP supported methods.'
      operationId: api_results_options_id_format
      parameters:
        - $ref: '#/components/parameters/resultId'
        - $ref: '#/components/parameters/formatParam'
      responses:
        '204':
          description: '`Allow` header _&lt;Response body is empty&gt;_'
          headers:
            Allow:
              $ref: '#/components/headers/Allow'
components:
  headers:
    ETag:
      description: Entity Tag (unique identifier provided for cache validation)
      schema: { type: string, pattern: "[0-9a-f]*" }
...
    Allow:
      description: List of HTTP supported methods
      schema: { type: string }
  schemas:
    User:
      type: object
      properties:
        id:
          description: 'User Id'
          type: integer
          format: int64
        email:
          description: 'User email'
          type: string
          format: email
        roles:
          description: 'User roles'
          type: string
          format: json
        _links:
          description: link relations
          type: object
          properties:
            parent:
              type: object
              properties: {"href": { "type": "string", "format": "url" }}
            self:
              type: object
              properties: {"href": { "type": "string", "format": "url" }}
      example:
        user:
          id: 1508
          email: 'miw.etsisi@upm.es'
          roles: [ 'ROLE_USER' ]
          _links:
            parent:
              href: /api/v1/users
            self:
              href: /api/v1/users/1508
    Result:
      type: object
      properties:
        id:
          description: 'Result Id'
          type: integer
          format: int64
        result:
          description: 'Result value'
          type: integer
          format: int64
        time:
          description: 'Result date in Y-m-d H:i:s format'
          type: string
          format: 'Y-m-d H:i:s'
        user:
          description: 'User who has scored'
          $ref: '#/components/schemas/User'
        _links:
          description: link relations
          type: object
          properties:
            parent:
              type: object
              properties: {"href": { "type": "string", "format": "url" }}
            self:
              type: object
              properties: {"href": { "type": "string", "format": "url" }}
      example:
        result:
          id: 1
          result: 111
          user:
            id: 1508
            email: 'miw.etsisi@upm.es'
            roles: [ 'ROLE_USER','ROLE_ADMIN' ]
            _links:
              parent:
                href: /api/v1/users
              self:
                href: /api/v1/users/1508
          time: '2025-01-01 10:10:10'
          _links:
            parent:
              href: /api/v1/results
            self:
              href: /api/v1/results/1
    Message:
      type: object
      description: HTTP response message
      properties:
        code:
          description: 'Response code'
          type: integer
          format: int64
        message:
          description: 'Response message'
          type: string
      example:
        code: 200
        message: Ok
  responses:
    304_NotModified_Response:
      description: "`NOT MODIFIED`"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-304:
              $ref: '#/components/examples/response-304'
        application/xml:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-304:
              $ref: '#/components/examples/response-304'
...
    401_Unauthorized_Response:
      description: '`Unauthorized`: invalid Authorization header'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-401:
              $ref: '#/components/examples/response-401'
        application/xml:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-401:
              $ref: '#/components/examples/response-401'
...
    403_Forbidden_NotOwn_Response:
      description: '`Forbidden`: The user does not own this Result or lacks permissions.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-403:
              $ref: '#/components/examples/response-403'
        application/xml:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-403:
              $ref: '#/components/examples/response-403'
    403_Forbidden_NotOwn_Response_json:
      description: '`Forbidden`: The user does not own this Result or lacks permissions.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-403:
              $ref: '#/components/examples/response-403'
...
    404_Resource_Not_Found_Response_Put_Result:
      description: '`Not found`: resource not found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-404-result:
              $ref: '#/components/examples/response-404-result'
            response-404-user:
              $ref: '#/components/examples/response-404-user'
        application/xml:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-404-result:
              $ref: '#/components/examples/response-404-result'
            response-404-user:
              $ref: '#/components/examples/response-404-user'
...
    404_Resource_Not_Found_Response:
      description: '`Not found`: resource not found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-404:
              $ref: '#/components/examples/response-404'
        application/xml:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-404:
              $ref: '#/components/examples/response-404'
...
    412_Precondition_Failed_Response:
      description: '`PRECONDITION FAILED`: one or more conditions given evaluated to false'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-412:
              $ref: '#/components/examples/response-412'
        application/xml:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-412:
              $ref: '#/components/examples/response-412'
...
    422_UnprocessableEntity_Response_Result:
      description: "`UNPROCESSABLE ENTITY`: Some request field does not have the correct format or some fields are left out"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-422:
              $ref: '#/components/examples/response-422-result'
        application/xml:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-422:
              $ref: '#/components/examples/response-422-result'
  requestBodies:
...
    ResultRequestBodyPut:
      description: "Request body for updating an existing Result"
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              result:
                description: "Result score (required, integer >= 0)"
                required: true
                type: integer
                format: int64
                minimum: 0
                example: 10
              time:
                description: "Optional time. Must follow 'YYYY-MM-DD HH:MM:SS' or similar. If omitted, no change"
                type: string
                format: "date-time"
                example: "2025-01-01 15:30:00"
              user:
                description: "Email of the user who owns this Result. Only ROLE_ADMIN can change to a different user; ROLE_USER must use their own email (or receive 403).It is not case sensitive"
                required: true
                type: string
                format: email
                example: "User_email@example.com"
          examples:
            updateResultExample:
              summary: "Minimal body for PUT"
              value:
                result: 20
                user: "User_email@example.com"
            updateResultWithTime:
              summary: "PUT example with time"
              value:
                result: 55
                time: "2026-05-10 09:15:00"
                user: "admin@miw.upm.es"
...
  parameters:
...
    resultId:
      name: resultId
      in: path
      description: 'ID of the Result (must be an integer).'
      required: true
      schema: { "type": "integer", pattern: '^\d+$', minimum: 1 }
    ifMatchParam:
      name: If-Match
      in: header
      description: "ETag value of the resource you are trying to update"
      required: true
      schema:
        type: string
        pattern: "[0-9a-f]*"
    ifNoneMatchParam:
      name: If-None-Match
      in: header
      description: "ETag value of the resource you are checking if it has been modified"
      required: false
      schema:
        type: string
        pattern: "[0-9a-f]*"
    formatParam:
      name: format
      in: path
      description: "Response format (json or xml)."
      required: true
      schema:
        type: string
        enum: ["json", "xml"]
        default: "json"
...
  examples:
    response-304:
      value:
        code: 304
        message: 'NOT MODIFIED'
...
    response-401:
      value:
        code: 401
        message: 'UNAUTHORIZED: invalid Authorization header'
    response-403:
      value:
        code: 403
        message: "FORBIDDEN You don't have permission to access"
    response-404:
      value:
        code: 404
        message: 'NOT FOUND: Resource not found'
...
    response-404-user:
      value:
        code: 404
        message: 'NOT FOUND: User User_email@example.com not found in db'
    response-404-result:
      value:
        code: 404
        message: 'NOT FOUND: Result with id 1 not found'
    response-412:
      value:
        code: 412
        message: 'PRECONDITION FAILED: one or more conditions given evaluated to false'
...
    response-422-result:
      value:
        code: 422
        message: 'UNPROCESSABLE ENTITY: Some request field does not have the correct format or some fields are left out'
  securitySchemes:
    MiWApiSecurity:
      type: http
      bearerFormat: JWT
      scheme: bearer